Fixed bug that made wrong players be assigned to own team
