+++ backend/core/tests/services/test_replays.py
@@ -1,0 +1,6 @@
+from django.test import TestCase
+
+
+class ReplaysTests(TestCase):
+    pass
+
