+++ backend/core/tests/services/test_round_replays.py
@@ -1,0 +1,6 @@
+from django.test import TestCase
+
+
+class RoundReplaysTests(TestCase):
+    pass
+
